requirements.txt
langchain-openai==0.0.2
requests==2.28.2
python-dotenv==1.0.0
streamlit==1.22.0